@@ -1,4 +1,4 @@
-name: MedicineReminder
+name: Vitals
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -34,6 +34,7 @@ dependencies:
   firebase_core: ^1.6.0
   firebase_auth: ^3.1.0
   google_sign_in: ^5.1.0
+  rxdart: ^0.27.2
 
 
   # The following adds the Cupertino Icons font to your application.
