@@ -1,4 +1,4 @@
-import 'package:MedicineReminder/main.dart';
+import 'package:Vitals/main.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
